Removed the underscore split from the importance network and plotting
Previously, the name of each node in importance network was
source + _ + layer. This was do distinguish between copies of nodes.
The underscore was removed by using str.split(). This caused problems
when there was an underscore in the name of the source. I refactored
it so that nodes now are defined by an ID which is unique for each source.
This way, there is no need to split.